Day 20.2: "Collision Detection" 🛬💣🚚
This commit dedicated to Unusual Sitting Positions.

It felt bad to splice apart my lazy sequenced `particle-path`
function. For this part, the evaluation chunking would have been
useless and hurt performance, so a simpler `next-particle` was
made, that does no lazy evaluation. I suppose `particle-path`
could be defined in terms of `iterate`, but it feels neat to
use `lazy-seq` myself, for now.